Some significant refactoring to our remoting layer and disable brittle unit tests until we are more stable

+++ src/main/org/jboss/messaging/core/server/ServerProducer.java
@@ -32,7 +32,7 @@ package org.jboss.messaging.core.server;
  */
 public interface ServerProducer
 {
-	long getID();
+	int getID();
 	
 	void close() throws Exception;
